Automatically restart container on deploy

+++ fabfile.py
@@ -37,8 +37,14 @@ def deploy():
     finally:
         fab.run('rm -rf %s' % tmpdir)
 
-    fab.sudo('install -o %s -d %s' % (fab.env.user, DEPLOY_DIR))
+    exists = fab.run('[ -d %s ] && echo 1 || echo 0' % DEPLOY_DIR).strip()
+    if exists == '0':
+        fab.sudo('install -o %s -d %s' % (fab.env.user, DEPLOY_DIR))
+    fab.run('docker stop powonline-frontend')
+    fab.run('docker rm powonline-frontend')
     fab.put('run-frontend.sh', '%s/run-frontend.sh.dist' % DEPLOY_DIR)
+    with fab.cd(DEPLOY_DIR):
+        fab.run('bash run-frontend.sh')
 
 
 @fab.task
